Created darts and towers and API for backend

See merge request compsci307_2020fall/final_team14!9

+++ src/ooga/backend/darts/SingleDart.java
@@ -2,7 +2,7 @@ package ooga.backend.darts;
 
 public class SingleDart extends Dart {
 
-  public SingleDart(int xPosition, int yPosition, int xVelocity, int yVelocity) {
+  public SingleDart(int xPosition, int yPosition, double xVelocity, double yVelocity) {
     super(xPosition, yPosition, xVelocity, yVelocity);
   }
 }
